datagen/pyproject.toml
[project]
name = "posttraining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=3.0.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "tenacity>=8.2.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
